Start of Server Backend (Gamelift, GameSessions, Player Sessions)
First we integrated GameLift with Unreal Engine. After we have setup use of the API Gateway through unreal engine API calls to enact AWS Lambdas for carrying out server functions for finding or creating a game sessions, making a player session, and after login, having the player session join the game session on the dedicated server. Right now we are starting to setup AWS Cognito for user verification and accounts.

Co-Authored-By: ethanfernand3s <163609112+ethanfernand3s@users.noreply.github.com>