Added Login, Signup, Confirmation, Dashboard Screen ,  Basic Player Functionality
By using a combination of AWS Cognito, Lambda, and Auth Tokens we were able to setup a secure sign in system that is carried out by AWS in a efficient and secure way.

After logging in the user is taken to a dashboard screen where they are able to view a few different screens before they join the game. After they do join the game they get taken to the Runeterra map. (ALL OF THESE SCREENS ARE WIP AND WILL HAVE ADDITIONIAL FUNCTIONALITY).

Everything next is just base functionality and is also WIP. But it is all written with replication in mind.

Added basic player functionality like moving, looking, jumping, crouching, opening game menu, and a basic attack.

Added base for weapons with weapon types based on the games classes and other customizable data for making each weapon. Also a weapon data assets for holding info on the weapons and animations.