Added Temp Assets For Assassin
EVERYTHING I ADDED IS EITHER WIP OR JUST FOR TESTING.

After creating a model on zbrush and the weapon on blender I used mixamo for some quick rigging and animation to bring in some mixamo animations for testing functionality. After which I setup a simple blendspace for the Assassin's direction/speed.